+++ src/adt/IAdtPackage.ts
@@ -32,11 +32,6 @@ export interface IUpdatePackageParams {
   record_changes?: boolean;
 }
 
-export interface IReadPackageParams {
-  package_name: string;
-  version?: 'active' | 'inactive';
-}
-
 export interface IDeletePackageParams {
   package_name: string;
   transport_request?: string;